Conflicts:
	timesheet/src/test/java/ut/org/catrobat/jira/timesheet/rest/UserRestTest.java

+++ timesheet/src/main/java/org/catrobat/jira/timesheet/activeobjects/Timesheet.java
@@ -63,7 +63,7 @@ public interface Timesheet extends Entity {
 
     int getEcts();
 
-    void setEcts(int ects);
+    void setEcts(double ects);
 
     String getLatestEntryDate();
